RabbitMQ

+++ depart6/src/main/java/com/depart/depart6/Dto/RabbitMessage.java
@@ -4,11 +4,11 @@ import java.io.Serializable;
 
 public class RabbitMessage implements Serializable {
 
-    private Long Id;
+    private int Id;
 
     private String name;
 
-    public RabbitMessage(Long id, String name) {
+    public RabbitMessage(int id, String name) {
         Id = id;
         this.name = name;
     }
@@ -16,11 +16,11 @@ public class RabbitMessage implements Serializable {
     public RabbitMessage() {
     }
 
-    public Long getId() {
+    public int getId() {
         return Id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         Id = id;
     }
